Added 3 alternate parts from Digikey with compatible footprints. Nice to have since the cheap parts from Korean Hroparts Elec sometimes are sold out
I might add that the Korean Hroparts Elec parts don't really wear that well. I've used them on prototype cards that I plug/unplug on a daily basis and they become "loose" over time. Makes me think that these are best for those things that are made to be pluggen in and left here.